requirements.txt
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
svgwrite==1.4
threadpoolctl==2.1.0
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
